+++ packages/cli/src/cli/react.ts
@@ -27,8 +27,9 @@ export class ReactCLI extends InlineCLI {
   ) {
     super(command, library, additionalModules);
 
-    this.program.hook('preAction', () => {
+    this.program.hook('preAction', (_thisCommand, actionCommand) => {
       if (this.program.opts().skipVersionCheck) return;
+      if (actionCommand.parent?.name() === 'git') return;
       checkMonorepoVersionConsistency([...REACT_LIBRARIES, Libraries.GT_I18N]);
     });
   }