+++ packages/sanity/CHANGELOG.md
@@ -1,5 +1,21 @@
 # gt-sanity
 
+## 2.1.0
+
+### Minor Changes
+
+- [#1872](https://github.com/generaltranslation/gt/pull/1872) [`1f55474`](https://github.com/generaltranslation/gt/commit/1f554742c76bee81b194c480a20cb5b697fafd6c) Thanks [@brian-lou](https://github.com/brian-lou)! - Add support for field-level localization
+
+### Patch Changes
+
+- [#1873](https://github.com/generaltranslation/gt/pull/1873) [`2e57a08`](https://github.com/generaltranslation/gt/commit/2e57a08a3f8539400f0aef5b90fff25c0a44dce1) Thanks [@ErnestM1234](https://github.com/ErnestM1234)! - Retrigger package publication after repairing release workspace dependency resolution.
+
+## 2.0.22
+
+### Patch Changes
+
+- [#1866](https://github.com/generaltranslation/gt/pull/1866) [`53e61ab`](https://github.com/generaltranslation/gt/commit/53e61abbb4522631f252bcd268eae9c079dd9967) Thanks [@ErnestM1234](https://github.com/ErnestM1234)! - Recover the stable releases skipped after the Odysseus version regression and publish the accumulated package changes under new npm versions.
+
 ## 2.0.18
 
 ### Patch Changes